requirements.txt
asgiref==3.5.2
bandit==1.7.4
black==22.8.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==4.1.1
django-crum==0.7.9
django-environ==0.9.0
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
filelock==3.8.0
flake8==5.0.4
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
identify==2.5.5
mccabe==0.7.0
mypy==0.981
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.1
pbr==5.10.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pytz==2022.2.1
PyYAML==6.0
smmap==5.0.0
sqlparse==0.4.3
stevedore==4.0.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
virtualenv==20.16.5
